Implements HDS-Forms part 1 - Add Column render view components - Add all resources - Generate Test data with HForm and FormModelFactory

+++ src/main/java/org/philimone/hds/forms/model/enums/ColumnType.java
@@ -3,7 +3,7 @@ package org.philimone.hds.forms.model.enums;
 public enum ColumnType {
 
     INTEGER,
-    DOUBLE,
+    DECIMAL,
     STRING,
     DATE,
     DATETIME,
